# openapi.yaml
openapi: 3.0.0
info:
  version: '1.1'
  title: Ruokavälitys API
  description: API for both, managing the RV system and it's inventory, and browsing and purchasing products.
...
servers:
  - url: https://rv-backend-dev.herokuapp.com/
    description: Staging environment
  - url: https://rv-backend.herokuapp.com/
    description: Production environment
components:
  securitySchemes:
    normalAccessToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description:
        A JWT access token used for authenticating with the \"end user facing\" portition of the API.
        Can be obtained with any valid credentials from route `/api/v1/authenticate`.
    adminAccessToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description:
        A JWT access token used for authenticating with the administrations portition of the API.
        Can be obtained with administrator credentials from route `/api/v1/admin/authenticate`.
  schemas:
...
    Category:
      type: object
      required: 
        - categoryId
        - description
      properties: 
        categoryId: 
          type: integer
          example: 10
        description: 
          type: string
          example: Drinks, soda with bottle return
    Preference:
      type: object
      required:
        - key
        - value
      properties:
        key:
          type: string
          example: globalDefaultMargin
        value:
          description: Can be any JSON value. The type depends on the preference in question.
          example: 0.05
  responses:
...
    NotAuthorizedErrorResponse:
      description: Your user lacks the required privilege level to perform this action.
      content:
        application/json:
          schema:
            type: object
            required:
              - error_code
              - message
            properties:
              error_code:
                type: string
                enum:
                  - not_authorized
              message:
                type: string
                example: Not authorized
    NotFoundErrorResponse:
      description: The resource your are requesting does not exist.
      content:
        application/json:
          schema:
            type: object
            required:
              - error_code
              - message
            properties:
              error_code:
                type: string
                enum:
                  - not_found
              message:
                type: string
                example: No product with barcode \'1312767817\' found
    BadRequestErrorResponse:
      description: The request you sent contains invalid values or is missing fields.
      content:
        application/json:
          schema:
            type: object
            required:
              - error_code
              - message
            properties:
              error_code:
                type: string
                enum:
                  - bad_request
                  - invalid_reference
              message:
                type: string
                example: Missing or invalid fields in request
              errors:
                type: array
                items:
                  type: string
                  example: 'boxBarcode should be a numeric 1-14 digit barcode' 
...
paths: 
...
  /api/v1/categories: 
    get: 
      summary: Get a list of product categories
      security:
        - normalAccessToken: []
      responses: 
        200: 
          description: Request fulfilled successfully
          content: 
            application/json: 
              schema: 
                type: object
                required: 
                  - categories
                properties: 
                  categories: 
                    type: array
                    items: 
                      '$ref': '#/components/schemas/Category'
        401:
          '$ref': '#/components/responses/NotAuthorizedErrorResponse'
  /api/v1/categories/{categoryId}: 
    get: 
      summary: Get details of a product category
      security:
        - normalAccessToken: []
      parameters: 
        - name: categoryId
          required: true
          in: path
          schema: 
            type: integer
      responses: 
        200:
          description: Request fulfilled successfully
          content: 
            application/json: 
              schema: 
                type: object
                required: 
                  - category
                properties: 
                  category: 
                    '$ref': '#/components/schemas/Category'
        401:
          '$ref': '#/components/responses/NotAuthorizedErrorResponse'
        404:
          '$ref': '#/components/responses/NotFoundErrorResponse'
...
  /api/v1/admin/categories: 
    get: 
      summary: Get a list of product categories
      security:
        - adminAccessToken: []
      responses: 
        200: 
          description: Request fulfilled successfully
          content: 
            application/json: 
              schema: 
                type: object
                required: 
                  - categories
                properties: 
                  categories: 
                    type: array
                    items: 
                      '$ref': '#/components/schemas/Category'
        401:
          '$ref': '#/components/responses/NotAuthorizedErrorResponse'
    post: 
      summary: Create a new product category
      security:
        - adminAccessToken: []
      responses: 
        201: 
          description: New product category created successfully
          content: 
            application/json: 
              schema: 
                type: object
                required: 
                  - category
                properties: 
                  category: 
                    '$ref': '#/components/schemas/Category'
        400:
          '$ref': '#/components/responses/BadRequestErrorResponse'
        401:
          '$ref': '#/components/responses/NotAuthorizedErrorResponse'
      requestBody: 
        required: true
        content: 
          application/json: 
            schema: 
              type: object
              required: 
                - description
              properties: 
                description: 
                  type: string
  /api/v1/admin/categories/{categoryId}: 
    get: 
      summary: Get the details of a product category
      security:
        - adminAccessToken: []
      parameters: 
        - name: categoryId
          required: true
          in: path
          schema: 
            type: integer
            example: 10
      responses: 
        200: 
          description: Request fulfilled successfully
          content: 
            application/json: 
              schema: 
                type: object
                required: 
                  - category
                properties: 
                  category: 
                    '$ref': '#/components/schemas/Category'
        401:
          '$ref': '#/components/responses/NotAuthorizedErrorResponse'
        404:
          '$ref': '#/components/responses/NotFoundErrorResponse'
    patch: 
      summary: Update details of a product category
      security:
        - adminAccessToken: []
      parameters: 
        - name: categoryId
          required: true
          in: path
          schema: 
            type: string
      responses: 
        200: 
          description: Category updated successfully
          content: 
            application/json: 
              schema: 
                type: object
                required: 
                  - category
                properties: 
                  category: 
                    '$ref': '#/components/schemas/Category'
        400:
          '$ref': '#/components/responses/BadRequestErrorResponse'
        401:
          '$ref': '#/components/responses/NotAuthorizedErrorResponse'
        404:
          '$ref': '#/components/responses/NotFoundErrorResponse'
      requestBody: 
        required: true
        content: 
          application/json: 
            schema: 
              type: object
              required: 
                - description
              properties: 
                description: 
                  type: string
    delete: 
      summary: Delete a product category
      security:
        - adminAccessToken: []
      parameters: 
        - name: categoryId
          required: true
          in: path
          schema: 
            type: string
      responses: 
        200: 
          description: Category deleted successfully
          content: 
            application/json: 
              schema: 
                type: object
                required: 
                  - deletedCategory
                  - movedProducts
                properties: 
                  deletedCategory: 
                    '$ref': '#/components/schemas/Category'
                  movedProducts: 
                    type: array
                    items: 
                      type: string
        401:
          '$ref': '#/components/responses/NotAuthorizedErrorResponse'
        403:
          description: 'Cannot delete the default product category'
          content:
            application/jsson:
              schema:
                type: object
                required:
                  - error_code
                  - message
                properties:
                  error_code:
                    type: string
                    enum:
                      - bad_request
                  message:
                    type: string
                    example: Cannot delete the default category
        404:
          '$ref': '#/components/responses/NotFoundErrorResponse'
...
  /api/v1/admin/preferences:
    get:
      summary: Get list of preferences and their values
      security:
        - adminAccessToken: []
      responses:
        200:
          description: Request fulfilled successfully
          content:
            application/json:
              schema:
                type: object
                required:
                  - preferences
                properties:
                  preferences:
                    type: array
                    items:
                      '$ref': '#/components/schemas/Preference'
        401:
          '$ref': '#/components/responses/NotAuthorizedErrorResponse'
  /api/v1/admin/preferences/{preferenceKey}:
    get:
      summary: Get the value of a preference
      security:
        - adminAccessToken: []
      parameters:
        - name: preferenceKey
          required: true
          in: path
          schema:
            type: string
            example: globalDefaultMargin
      responses:
        200:
          description: Request fulfilled successfully
          content:
            application/json:
              schema:
                type: object
                required:
                  - preference
                properties:
                  preference:
                    '$ref': '#/components/schemas/Preference'
        401:
          '$ref': '#/components/responses/NotAuthorizedErrorResponse'
        404:
          '$ref': '#/components/responses/NotFoundErrorResponse'
    patch:
      summary: Update the value of a preference
      security:
        - adminAccessToken: []
      parameters:
        - name: preferenceKey
          required: true
          in: path
          schema:
            type: string
            example: globalDefaultMargin
      responses:
        200:
          description: Preference's value updated successfully
          content:
            application/json:
              schema:
                type: object
                required:
                  - preference
                properties:
                  preference:
                    '$ref': '#/components/schemas/Preference'
        400:
          '$ref': '#/components/responses/BadRequestErrorResponse'
        401:
          '$ref': '#/components/responses/NotAuthorizedErrorResponse'
        404:
          '$ref': '#/components/responses/NotFoundErrorResponse'